+++ scripts/create2/xerc20-zai-remote.ts
@@ -8,7 +8,7 @@ async function main() {
 
   const implArgs = [
     "ZAI Stablecoin",
-    "USDz",
+    "xUSDz",
     "0x1f09ec21d7fd0a21879b919bf0f9c46e6b85ca8b", // admin
   ];
 
@@ -18,9 +18,9 @@ async function main() {
     "XERC20",
     implArgs,
     get("XERC20-impl", network.name),
-    "arbitrum",
-    "0x07f7b4a675919f14b4ac361f5bbd083b8493fd1b3d43b957013f2eb971273abc",
-    "0x69000cc63be9b4322f3f62c233dd1a7f509ae080"
+    network.name,
+    "0x2fcf32741f80f7e50bab7728ed54fd9f51edf766b3dda12113c84096ffc2fae8",
+    "0x69000a259d21c1824c7ffe3fd1e19a557a81d87e"
   );
 }