+++ tasks/connext/test-restake.ts
@@ -8,15 +8,15 @@ task(`test-restake`).setAction(async (_, hre) => {
   const conifgLocal = config[hre.network.name];
   assert(!!conifgLocal, `Config not found for ${hre.network.name}`);
 
-  const contractD = await hre.deployments.get("L2DepositCollateral");
+  const contractD = await hre.deployments.get("L2DepositCollateralConnext");
 
   const contract = await hre.ethers.getContractAt(
-    "L2DepositCollateral",
+    "L2DepositCollateralConnext",
     contractD.address
   );
   const erc20 = await hre.ethers.getContractAt("MockERC20", conifgLocal.usdc);
 
   console.log("deposit contract is at", contract.target);
   await waitForTx(await erc20.approve(contract.target, MaxUint256));
-  await waitForTx(await contract.deposit(10e6, 0, Date.now()));
+  await waitForTx(await contract.deposit(1e6, 0, Date.now()));
 });